+++ spec/models/maths_and_physics/dqt_record_spec.rb
@@ -9,44 +9,44 @@ RSpec.describe MathsAndPhysics::DQTRecord do
   describe "#eligible?" do
     EXAMPLE_ELIGIBLE_JAC_CODES.each do |jac_code|
       context "when the given ITT subject (#{jac_code}) is eligible" do
-        let(:attributes) { {"ITTSub1Value" => jac_code} }
+        let(:attributes) { {itt_subject_jac_codes: [jac_code], degree_jac_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "19/3/2017"})).eligible?). to eql true
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?). to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "1/10/2015"})).eligible?). to eql true
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?). to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "8/3/2000"})).eligible?). to eql false
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?). to eql false
         end
       end
 
       context "when the given degree (#{jac_code}) is eligible" do
-        let(:attributes) { {"HESubject1Value" => jac_code} }
+        let(:attributes) { {degree_jac_codes: [jac_code], itt_subject_jac_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "19/3/2017"})).eligible?). to eql true
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?). to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "1/10/2015"})).eligible?). to eql true
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?). to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "8/3/2000"})).eligible?). to eql false
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?). to eql false
         end
       end
     end
 
     EXAMPLE_NON_ELIGIBLE_JAC_CODES.each do |jac_code|
       context "when the given ITT subject or degree (#{jac_code}) isn't eligible" do
-        let(:attributes) { {"ITTSub1Value" => jac_code, "HESubject1Value" => jac_code} }
+        let(:attributes) { {itt_subject_jac_codes: [jac_code], degree_jac_codes: [jac_code]} }
 
         it "always returns false" do
-          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({"dfeta qtsdate" => "19/3/2017"})).eligible?). to eql false
+          expect(MathsAndPhysics::DQTRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?). to eql false
         end
       end
     end