+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog]
 ## [Unreleased]
 
 - Exclude claims from previous academic years on matching claims screen
+- Fix a bug with automated DQT checks where the report has multiple rows for the
+  same claim
 
 ## [Release 066] - 2020-03-25